% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMarkers.R
\name{filterMarkers}
\alias{filterMarkers}
\title{remove markers with either all no-call or no variation
also generates default error and dropout rate inputs for biallelic SNPs based
on what is typically used as input for SNPPIT}
\usage{
filterMarkers(baseline, mixture = NULL, unsamp = NULL)
}
\arguments{
\item{baseline}{dataframe input for baseline}

\item{mixture}{dataframe input for mixture}

\item{unsamp}{dataframe input for unsampledPops}
}
\description{
remove markers with either all no-call or no variation
also generates default error and dropout rate inputs for biallelic SNPs based
on what is typically used as input for SNPPIT
}
