% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDFGEN_to_GmaBaseline.R
\name{IDFGEN_to_GmaBaseline}
\alias{IDFGEN_to_GmaBaseline}
\title{create the baseline dataframe needed by createGmaInput from IDFGEN objects}
\usage{
IDFGEN_to_GmaBaseline(baselinePops, Markers)
}
\arguments{
\item{baselinePops}{either a PopList or a Population}

\item{Markers}{the marker names (SNPs only) to use, in order of how they are used}
}
\description{
create the baseline dataframe needed by createGmaInput from IDFGEN objects
}
