% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDFGEN_to_GmaMixture.R
\name{IDFGEN_to_GmaMixture}
\alias{IDFGEN_to_GmaMixture}
\title{create the mixture dataframe needed by createGmaInput from IDFGEN objects}
\usage{
IDFGEN_to_GmaMixture(mixturePops, Markers)
}
\arguments{
\item{mixturePops}{either a PopList or a Population}

\item{Markers}{the marker names (SNPs only) to use, in order of how they are used}
}
\description{
create the mixture dataframe needed by createGmaInput from IDFGEN objects
}
